refactor response files
